refactor : 카테고리를 호스트로 변경 시에 선택하도록 리팩토링
